Default Commit Message

+++ AgentX/src/agentx/model/Boss.java
@@ -17,18 +17,18 @@ public class Boss implements Serializable {
     private double damage;
     private int y;
     private int x;
-    private int health;
+    private double health;
 
     public Boss() {
     }
 
-    public Boss(String bossName, double damage, int health) {
+    public Boss(String bossName, double damage, double health) {
         this.bossName = bossName;
         this.damage = damage;
         this.health = health;
     }
 
-    public Boss(String bossName, double damage, int y, int x, int health) {
+    public Boss(String bossName, double damage, int y, int x, double health) {
         this.bossName = bossName;
         this.damage = damage;
         this.y = y;
@@ -77,11 +77,11 @@ public class Boss implements Serializable {
         this.damage = damage;
     }
 
-    public int getHealth() {
+    public double getHealth() {
         return health;
     }
 
-    public void setHealth(int health) {
+    public void setHealth(double health) {
         this.health = health;
     }
